+++ src/standard/pro_que_builder.rs
@@ -73,9 +73,9 @@ impl ProQueBuilder {
             },
         };
 
-        let device = context.get_device_by_index(self.device_idx);
+        let device = context.get_device_by_wrapping_index(self.device_idx);
 
-        let queue = try!(Queue::new(&context, Some(device.clone())));
+        let queue = try!(Queue::new(&context, device));
 
         let program = try!(Program::from_parts(
             try!(program_builder.get_src_strings().map_err(|e| e.to_string())), 
@@ -121,6 +121,18 @@ impl ProQueBuilder {
         self
     }
 
+    // /// Sets a device index to be used and returns the `ProQueBuilder`.
+    // ///
+    // /// Defaults to `0`, the first available.
+    // ///
+    // /// This index WILL round robin, in other words, it cannot be invalid.
+    // /// If you need to guarantee a certain device, create your parts without
+    // /// using this builder and just call `ProQue::new` directly.
+    // pub fn device<'p>(&'p mut self, device: Device) -> &'p mut ProQueBuilder {
+    //     self.device_idx = device_idx;
+    //     self
+    // }
+
     /// Adds some source code to be compiled and returns the `ProQueBuilder`.
     ///
     /// Creates a `ProgramBuilder` if one has not already been added. Attempts
@@ -130,14 +142,21 @@ impl ProQueBuilder {
     /// source files. Pass an *unbuilt* `ProgramBuilder` to the 
     /// `::program_builder` method (described below).
     pub fn src<'p, S: Into<String>>(&'p mut self, src: S) -> &'p mut ProQueBuilder {
-        match self.program_builder {
-            Some(ref mut program_builder) => { program_builder.src(src); },
-            None => self.program_builder = {
-                let mut pb = Program::builder();
-                pb.src(src);
-                Some(pb)
-            },
-        };
+        // match self.program_builder {
+        //     Some(ref mut program_builder) => { program_builder.src(src); },
+        //     None => self.program_builder = {
+        //         let mut pb = Program::builder();
+        //         pb.src(src);
+        //         Some(pb)
+        //     },
+        // };
+        if self.program_builder.is_some() {
+            panic!("ocl::ProQueBuilder::src: Cannot set src if a 'ProgramBuilder' is already \
+                defined. Please use the '::program_builder' method for more complex build \
+                configurations.");
+        } else {
+            self.program_builder = Some(Program::builder().src(src))
+        }
 
         self
     }
@@ -152,12 +171,12 @@ impl ProQueBuilder {
     /// `::device_idxs` method). `ProQueBuilder` will only build programs for
     /// the device specified by `::device_idx` or the default device if none has
     /// been specified.
-    pub fn program_builder<'p>(&'p mut self, program_builder: ProgramBuilder) -> &'p mut ProQueBuilder {
-        assert!(self.program_builder.is_none(), "ProQueBuilder::program_builder(): Cannot set the \
+    pub fn prog_bldr<'p>(&'p mut self, program_builder: ProgramBuilder) -> &'p mut ProQueBuilder {
+        assert!(self.program_builder.is_none(), "ProQueBuilder::prog_bldr(): Cannot set the \
             'ProgramBuilder' using this method after one has already been set or after '::src' has \
             been called.");
 
-        assert!(program_builder.get_devices().len() == 0, "ProQueBuilder::program_builder(): The \
+        assert!(program_builder.get_devices().len() == 0, "ProQueBuilder::prog_bldr(): The \
             'ProgramBuilder' passed may not have any device indexes set as they will be unused. \
             See 'ProQueBuilder' documentation for more information.");